Fixed Tests, Run Test like normal Django Apps
